+++ yalafi/packages/graphicx.py
@@ -1,0 +1,24 @@
+
+#
+#   YaLafi module for LaTeX package graphicx
+#
+
+from yalafi.defs import ModParm
+
+require_packages = []
+
+def modify_parameters(parms):
+
+    macros_latex = r"""
+
+        \newcommand{\includegraphics}[2][]{}
+
+    """
+
+    macros_python = []
+
+    environments = []
+
+    return ModParm(macros_latex=macros_latex, macros_python=macros_python,
+                        environments=environments)
+