+++ yalafi/packages/xcolor.py
@@ -3,18 +3,29 @@
 #   YaLafi module for LaTeX package xcolor
 #
 
-from yalafi.defs import ModParm
+from yalafi.defs import Macro, ModParm
 
 require_packages = []
 
 def modify_parameters(parms):
 
     macros_latex = r"""
-        \newcommand{\color}[1]{}
-        \newcommand{\colorbox}[2]{#2}
-        \newcommand{\fcolorbox}[3]{#3}
-        \newcommand{\textcolor}[2]{#2}
+
+        \newcommand{\color}[2][]{}
+        \newcommand{\colorbox}[3][]{#3}
+        \newcommand{\definecolor}[4][]{}
+        \newcommand{\textcolor}[3][]{#3}
+
     """
 
-    return ModParm(macros_latex=macros_latex)
+    macros_python = [
+
+        Macro(parms, '\\fcolorbox', args='OAOAA', repl='#5'),
+
+    ]
+
+    environments = []
+
+    return ModParm(macros_latex=macros_latex, macros_python=macros_python,
+                        environments=environments)