+++ HISTORY.md
@@ -1,3 +1,10 @@
+Work in progress
+----------------
+- yalafi
+    - added extension modules
+    - added tests for extension modules
+- added file list-of-macros.md
+
 Version 1.1.2 (2020/05/28)
 --------------------------
 - yalafi